Initial vec3 size cleanup

+++ app/src/main/cpp/helper.cpp
@@ -54,7 +54,7 @@ char *loadFileToMemory(AAssetManager *mgr, const char *filename) {
     return fileContent;
 }
 
-void createScalar3DTexture(GLuint *id, vec3 size, float* data){
+void createScalar3DTexture(GLuint *id, ivec3 size, float* data){
 
     glGenTextures(1, id);
     glBindTexture(GL_TEXTURE_3D, *id);
@@ -66,7 +66,7 @@ void createScalar3DTexture(GLuint *id, vec3 size, float* data){
     glTexParameteri(GL_TEXTURE_3D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
 }
 
-void createVector3DTexture(GLuint *id, vec3 size, vec3* data){
+void createVector3DTexture(GLuint *id, ivec3 size, vec3* data){
 
     glGenTextures(1, id);
     glBindTexture(GL_TEXTURE_3D, *id);  // todo RGB16F is not considered color-renderable in the gles 3.2 specification. Consider switching to RGBA16F
